Moved homepage content to about page.
There is useful information on the homepage, however that realestate is going to be
more valuable as a project overview dashboard of some kind. In order to preserve the
content that was there, it was moved onto the about page, and the about page has
had its lorem ipsum content removed. I've also added a contribution block in the side
bar for the project.

Change-Id: I6d7b8a5c4a1252df3b19139ba1f4bd73805529d8