Simple round trip API integration with /v1/teams
This commit makes use of previous API inclusions and resource frameworks to
demonstrate a simple list round-trip from server to client on the teams endpoint.
To properly see it in action, you'll have to manually add a team to your local
database instance.

Further development will depend on a few minor changes to our API layer, as it
currently uses strings as keys, which isn't really RESTful.

Change-Id: Ie95685a6fd3cd3ab2b674bef3685b2896eb72f0d