Simple round trip API integration with storyboard-api
This commit makes use of previous API inclusions and resource frameworks to
demonstrate a simple list round-trip from server to client. To properly see
it in action, please retrieve CR 68540 and run that alongside grunt server.
It will allow you to create, edit, and list project groups with a simple UI.

Change-Id: Ie95685a6fd3cd3ab2b674bef3685b2896eb72f0d