Updated grunt proxy server to rewrite to WSGI in dev
When running storyboard as a developer, we run into a bit of a problem with
the apache mod_wsgi config (which prefixes a context root of /api/v1) vs.
our simple-server setup (which uses /v1 as a context root). Since this only
impacts the grunt server running locally, it seems appropriate to simply
rewrite the grunt CORS-buster proxy to handle this for us.

Change-Id: Iec9454c845a880f383a8c3a7bfc15db58101d549