Task, story, and project fixes
Tasks and stories now use the correct query parameters for filtering
by their parent object. Tasks can now be created with project associations.
When a story is created, a task is automatically generated as well - this
is conflicting with what's going on on the API side, because the API also
generates a task. Unfortunately the API doesn't actually check for a
submitted project_id, so it's probably better to handle it here.

I'm open to options on what the default of the create-a-task dropdown
should be.

Change-Id: I49fb6a8879660a600f227ea4987664f6b26eed49