Fix NPM CA Chain issue and port conflict.
This patch includes two fixes. First, it disables our functional tests
because of a bug in grunt-connect-proxy, which is fixed by an
as-yet-not-accepted patch here:

https://github.com/krotscheck/grunt-connect-proxy/commit/81631b20c537e1d07071c07299b9f040671737d2

Secondly, it includes a temporary fix for the recent NPM CA chain update
explained here:

http://blog.npmjs.org/post/78085451721/npms-self-signed-certificate-is-no-more

The basic premise is that the node package manager no longer supports
their self-signed certificate. This is good! Unfortunately, the
released version of npm is 1.4.4, but the one packaged with nodejs
0.10.26(stable) is 1.4.3. Until nodejs updates itself, we need to
temporarily bork the ca chain to update npm.

Since we're using throwaway slaves, and it's a temporary fix, and
node is only used for builds, I feel this is an acceptable risk.

Change-Id: Ic510e55355966591affbe25ce3f831ca424b2afa