Fixed NPM CA chain issue
This is a temporary fix due to a version mismatch on the current
published nodejs code. The issue is highlighted here:

http://blog.npmjs.org/post/78085451721/npms-self-signed-certificate-is-no-more

The basic premise is that the node package manager no longer supports
their self-signed certificate. This is good! Unfortunately, the
released version of npm is 1.4.4, but the one packaged with nodejs
0.10.26(stable) is 1.4.3. Until nodejs updates itself, we need to
temporarily bork the ca chain to update npm.

Since we're using throwaway slaves, and it's a temporary fix, and
node is only used for builds, I feel this is an acceptable risk.

Change-Id: Ic510e55355966591affbe25ce3f831ca424b2afa