Create story modal allows multi-task creation.
The story modal now allows the creation of multiple tasks in one
use action. When only one task is active, and it has not been
modified, the title is slaved to the title of the story, however
as soon as the task title is modified, it becomes independently
editable. In addition, projects are independently selectable
per task.

Change-Id: If1d8e06b2a79708d117e805788854e8da0e1933f