Creating a new story now requires authentication.
The UI should not allow an unauthenticated user to create stories;
recording the identity of a reporter will allow us to communicate with them
to discover specifics about their requests. This commit adds a login modal
and controller in the auth package, which is hooked into the new story modal.

Change-Id: I62e6937f863c8d0f790d7bff7edd7335b9ac4653