Added UI switches for authentication state.
When a user is not authenticated, many UI Controls should be hidden,
including "add" buttons, "edit" buttons, "delete" buttons. Furthermore,
during the transition of the session state (when someone logs in/out),
the UI routes need to re-resolve themselves to determine whether the user
is now on an invalid state.

Change-Id: I057c5e5db30a74edfe66c67b60d2c0b797da6561