Added pagination directives
I'm injecting a global pagesize parameter for the UI to use across
all services that adhere to the storyboard contract. Then, I set
that as a default parameter within our API signature. This is
essentially a shim to ensure that for now, all of our content shows
up when we ask for it. More sophisticated paging will be required
eventually.

Also, removed unused header directives, and corrected the count
in the projects and stories list page.

Depends on https://review.openstack.org/#/c/83568

Change-Id: I4cb08e93ac57d089e6201236c3f640f0b2f9a31f