Fixed contenteditable
While sexy, contenteditable is poorly implemented across browsers, particulary
between chrome and firefox. Rather than try to write complex javascript to accomodate
for all edge cases, I removed it and replaced it with a more traditional input field.
I also added a .context-edit class underneat the form-control group that allows us
to conceal the border of the input control while unfocused. I also added a series
of type-style annotation classes that allow an input field to imitate a header of
the same name.

Change-Id: I61d7afbffcc48f014dff47ee7fbfbbe9f8d46311