Tasks are now editable
Added a simple edit form to the task list on the story page. Again: Not
pretty, but I want to get all of the functionality working before I
engage with the UX guys and get them to do a once over on this.

- New inline templates for the edit form.
- New control methods for the list item controller.
- Logged in stylesheet, for mouseover highlighting of editable
components, but only when logged in.
- Task delete modal.
- Updated task status dropdown so that its click events don't pollute
the rest of the DOM.

Change-Id: I9d3bb8c19cc0a0b7fcec23ce2637996f99a9a3b3