Added task status display to story list view
In combination with the aggregate task statuses coming back from the API,
this change surfaces those in the story list view. It also cleans up the
UI a bit, displaying only title, #, and aggregate status. Expanding the
row provides more details about the story's current state, as well as
revealing the description.

Change-Id: I28c2475f3322aa97d285c38b9b45576ca7c11ccc