Fixed display of user-entered whitespace.
When entering descriptions with carriage returns, the UI would not
properly display them, collapsing linefeeds into one big paragraph.
This fix introduces the 'honor-carriage-return' style which can
be applied to any text element to get it to support both tabs and
carriage returns. A note on the wacky markup: ANYTHING within the
tag is formatted as is, including code indentations and more. In
other words, we need to make sure that the data binding is not
surrounded by any.

Change-Id: Ie858a33003b1b0ae6883d9d3768f903d2ee3f9b1