Moved main navigation into fixed sidebar.
The more I used it and tried to design for it, the more the bootstrappy
header and footer menus started to annoy me. This commit removes them,
and moves primary navigation into a left-locked sidebar. This reclaims
some of the bootstrap wasted whitespace, as well as giving us an
option for two tiers of submenu: directly below the main menu item,
or pulled to the right as a second-tier left menu.

- Removed header and footer as no longer necessary.
- Moved all links into main_menu and restyled.
- Deleted old zombie bootstrap files, not certain where they came from.
- Removed old navbar styles.
- Added new navbar-fixed-left component, as well as a pilltab and
link menu style.
- Added margin and container adjustments for responsive sizing.
- Updated body padding since we don't have top fixed menus anymore.
- Updated global container widths to accomodate for 150px less space.
- Removed zombie code in index.html that should never have been there.
- Updated Bootstrap

Change-Id: Ib84ba77ea296e7ad0c6e8482c7b1c3a31885eb03