Enabled functional tests
An outstanding bug in grunt-connect-proxy v0.1.8 caused an error
in our selenium UI tests. I submitted a patch that was accepted
in v0.1.9, however that version was never properly deployed to
NPM due to the recent CA cert chain changes. The package
maintainer finally released a new version.

This commit upgrades the proxy to v0.1.10 and re-enables our
selenium tests.

Change-Id: Ia944f59e8b6f303a956ca2dbe7b506cf231c79bd