Favicon Whitelabel
This change allows us to whitelabel the favicon in the browser. Much
like the simple-whitelabel patch already landed, this makes use of
array ordering and file resolution to determine which favicon 'wins'.
In this case, both icons are copied, however if the custom font icon
actually exists, it will override storyboard's default icon.

Change-Id: Ic412a920d034b965ced25a7b63a51f82c4447a3d