Fixed padding on refresh icon.
When logging in, the refresh icon was glued up against the top of
the screen without any padding. That's a sideffect of moving things
to a sidebar that I hadn't anticipated. I've added some padding. I
also altered the icon used to the refresh arrows that we use
everywhere else, and used a gray font because the black
was a bit stark against the white screen.

Change-Id: I661dc6eda52f32049cbe052c8ead80431362c6c5