Autoassign user when state moves into progress
In order to keep somewhat better track of who's working on what,
I've added an automatic update switch that will assign a task to
the current user if the state has been updated to one that indicates
that work is being done, but ONLY if the task is not already assigned.

I feel that it's better to not enforce this on the API layer, because
we may end up with quite a few tasks assigned to jenkins.

Change-Id: I7e68fce7d6c5b5a1aa28d7382a0b870796d80993