Added resolveSessionState method
I added a new resolver method to our session resolver, which allows
us to determine whether a user is logged in or not before we finish
resolving the session. As a result we can actually inject the current
session state in our state controller, without having to write
additional complex code to determine what it is.

Change-Id: I90dde57679cca24cc3da6ea67f43276c984fa759