Added simple logged-in dashboard
In order to make tasks assigned to individuals more usable, I've added
a separate logged-in dashboard view that provides us with a list of
all stories that contain a task assigned to the current user. I've
also updated the sidebar menu to be labeled 'Dashboard', and added
a directive with which we can centralize our layout for story task
status.

Change-Id: I92c39cf3bae26424ef07f753feebaadb381d0739