Textareas now autoresize their height.
I added the monospaced.elastic library from bower, as it already
implements everything we need for the textarea resize. Then I applied
it to several of the textareas throughout the application. In a
few cases I had to add a small timeout trigger to force a recalculation
of the text area, because the initial size check was performed while
the element in question was hidden which confuses browsers.

Change-Id: I368096c93f61142df472573e84bac2d13b4647f0