Added search icon to typeahead fields
The typeahead search fields needed some kind of icon to indicate that
they weren't just plain text fields. This adds the search icon as a
muted out option on the right. I also switched the CSS style and the
made the line-height dependent on variables, and pulled the task_row
template into its own (inline) template (mostly because of the
character row limit).

Change-Id: I973190779bd654d24125d977233c895407c67a29