Refresh token support
A refresh manager added.

The access token will be checked every 30 minutes for expiration. If it
is expired or will expire soon a new acces token will be requesd. The
expiration is also checked when an access token is validated at Session
init state. So when a user opens a tab with Sotoryboard an expired token
will be refreshed.

Change-Id: I8ce8f8800a36785cf761c1dbb486c58686a25498