New StoryBoard Logo
This patch adds both the openstack and storyboard font glyph source SVG
files to our repostiory, as well as a tool that allows us to convert
those SVG files into a font (with the associated LESS). The tool at
the moment still has to be run manually, because grunt-webtool requires
fontforge which is not installable via NPM (you can run it manually
using 'grunt webfont').

The old custom font has been removed.

(Put up as a separate patch because it's likely to be contentious)

Change-Id: I60be7ea3b8fce7cc210538719cc1ec5d20d50f60