New StoryBoard Logo
This patch adds both the openstack and storyboard font glyph source SVG
files to our repostiory, as well as a tool that allows us to convert
those SVG files into a font (with the associated LESS). It's integrated
into our build, and listens to /fonts/src from which the custom font
files and css are generated.

The old custom font has been removed.

(Put up as a separate patch because it's likely to be contentious)

Change-Id: I60be7ea3b8fce7cc210538719cc1ec5d20d50f60