Error message & notification handling
This patch adds the notification module, a mostly-self contained way
of surfacing errors, warnings and notifications to the user. These
can either be initiated by the application, or can be received from
the server. Among other things, the impact of this will be that
the user will be notified if one of our HTTP requests fails. This
fixes the situation where storyboard would 'fail silently' when
someone is trying to accomplish something.

- New notification module.
- Notifications directive, for a list of all notifications that
made it past filters.
- New Notification service that handles all error broadcasts in the
system.
- New severity directive.
- HTTP Error broadcaster now uses the notification service.
- Session now uses Notification service to handle 401 errors.
- Added new notification filters to strip out 200's and template load
requests.

Change-Id: I4aaa50404560d1c1f14f639e3ac68b2ac2d4380c