Added custom CSS aliases for common storyboard icons.
We're starting to use iconography within our UI, so it makes sense
to centralize the icons we're using for different resource types.
To accomplish this, I've added fontawesome to our LESS manifest,
removed it from our index.html (since now it's compiled into our
central theme), and created CSS class aliases for the common icons that
we use.

Change-Id: I7c9f5e1791a6152e3fdaea95dd80550b941f5d05