Renamed module folders for consistency
This is a rename-only change. There's inconsistent naming of
folders in the various javascript subdirectories, i.e. whether
the directory that houses controllers should be called 'controller'
or 'controllers'. I've consolidated all names to use the singular
version, though that choice was mainly made because there
were more folders using that convention.

Change-Id: If098634495355c85399cbc265524c8d89b45b09b