Tag Complete directive has proper focus behavior
The wrapper around the tag-complete method has been taught to delegate
its focus method to the encapsulated <input>, and an additional class
was added to apply the bootstrap-highlight-focus style to the control.
Note that to properly respond to tab focus, we have to explicitly issue
a $digest() after both focus(added) and blur(already there).

Change-Id: I488140abeebaa2b0d8d484058f2f96590cb8ca24