Subscription
This module allows a user to subscribe to a resource. This module does
not include reading an events stream of changes to subscribed resources.
Provided are the following:

- A subscription directive that is configurable by resource and ID. It
is sensitive to the logged-in state and the current ID, and will resolve
its subscription state accordingly.
- A service which allows a user to read and write their
subscriptions.
- Subscription UI added to various resource list items.

Depends on https://review.openstack.org/108516

Change-Id: I2e9181e93209ce519f0a5eeb61630b1724a2527c