Removed nodeenv dependency
Once #106152 lands, nodejs will be available on our trusty build nodes,
so nodeenv is no longer strictly required. However, not everyone wants
to install nodejs on their system, so we now wrap the nodeenv
invocation into a script to test whether a local node exists. Since
we can no longer assume the existence of a clean virtual environment,
we also no longer install a global version of bower and grunt, and
instead use the versions bundled with the codebase's package.json.

Change-Id: I42f4b61037b8e988708b39e951dd2cddd99313b0