Added pagesize to browse service.
The default page size of 500 is causing a bit of a slowdown
on storyboard.o.o when using the typeahead in search and
quicknav. This patch applies a few changes to the Criteria and the
ResourceFactory as follows:

- Default pagesize for criteria searches is now Preference.page_size.
- PageSize can be overridden in the actual invocation of the
criteria resolver and/or the search builder.
- SearchResults have been fixed to use the user's preference setting
instead of a hardcoded 20.
- List population for the typeahead dropdowns has been hardcoded to
no more than 5 records per type, since the speed of a user's typing
tends to alter the result set very quickly.

Change-Id: I12c4fa8d62e0510ebd2287a709d93d6f922fdfcd
Story: 131