Quicknav is now even quicker!
This patch adds a search-string precheck that will route the
quicksearch result queries in one of two ways: If it's entirely
numeric, it assumes that you're trying to navigate by unique ID,
in which case it will try to present you options with that unique
id. If it is not entirely numeric, it will fall back to the LIKE
text search on the primary title.

Change-Id: I823c509196d21c587370351583abf87b950ee2a9