Autoconfiguration before bootstrap.
This patch delays the application bootstrap until we've tried to load
and inject parameters saved in an external configuration file. If that
file isn't there, the application will fall back to its own sane
defaults, however this will give us an opportunity to more gracefully
configure the webclient for each install. Example: Client and API are
not on the same host.

Change-Id: I1d0ce573dd05faa1efb325871ad5c30f08529205