Allow assigning tasks without expanding the edition box
Add an autocomplete box for assignee instead of the fixed label
that was now. Allow selecting user from there, and updating
automatically to In progress, with the same behaviour as the
assignee edition field.

Autocomplete box will be replacing a clickable label,
so we don't break the layout and allow inline edition.

Change-Id: Ifac7905e28495550599dd0532b928816d23e3192