Dynamically refresh events on task changes
Changes done on tasks are causing events to be created. Currently
new generated events were not visible until story detail page
was reloaded. Add an event refreshEvent that is triggered on every
change affecting to tasks, and collect this event in the proper
timeline controller, to reload the events dynamically.

Change-Id: I42ee71f59a9f85c4b8ecfa3d09e4943c69d6cfe6