Gruntfile Connect Hostname commandline option
The default hostname in grunt connect plugin may now be set using the
'--hostname VALUE' option. This allows a developer to bind to alternate
IP addresses such as 0.0.0.0, making the grunt server accessible from a
host machine if doing development on a VM with port forwarding.

Documentation will be updated in a separate commit on
openstack-infra/storyboard

Change-Id: I94d48b1be5bb22e05d8f574e71d6886b21f23daa