Eabled custom dev configuration.
Sometimes it is useful to allow an engineer to point the webclient
at an API not running on localhost. To address this edge case,
I have added src/config.json to .gitignore, and taught grunt how
to copy it during the dev build/deploy phase. If the file is not
found, it is ignored.

Do not merge until an equivalent patch to storyboard/docs is
provided.

Change-Id: I82229f92271d9094ec8ad4b52c38f3bc7ef7fa42