Read user preferences from backend
Remove usage of localStorage to retrieve preferences,
and add a Preference resource that uses the backend API.
Add methods in CurrentUser to retrieve and cache user
preferences once logged, and take care of updating prefs
when needed.
Update the prefs saving methods to save the settings using
the API.
Update the search methods to retrieve preferences from cache.

Change-Id: Icf3072e24fef86848801a5c72f5833605b1afcd7