Add timeout to the blur event for story inline edition
Currently story assignee and projects were only selectable
using keyboard, using mouse it was failing because the event
of hiding form was interacting with the mouse event. Add
a timeout in case of hiding form for that two typeahead
components to prevent that failure.
Also remove a duplicate method call from onBlur and onSelect
leaving only that on typeahead select.

Closes-Bug: 2000065

Change-Id: Ibb66d31ba0aa0228a0b63e2ff41ca7f2aa71d513