Fix in project groups validation
Detected several issues: we weren't validating minimum
lengths and regex pattern. Also there was a bug on adding
a new project group. It was possible to send new project
groups without initial project.
This seems to be a bug in Angular, when there is an inline
form with only typeahead, it's not setting the associated
form to dirty and just skips validation. Add the checkValidProjects
method as a workaround.

Change-Id: Id1938fcf6c7eea9475def51396961def3ba6f6cc