Added root state
A root state allows us to resolve needed application data before any
routes are resolved. It required that I prefix all of our states with
a 'sb' so that the statemachine blocks during resolution, however this
will benefit us as we layer on user preferences and similar
functionality. At this point, we are resolving the current session
state, which will refresh all of our oauth tokens before the route
concludes.

Change-Id: I4b315b9309db8eb296c27f0fb75b3d11854f91fc