Don't add unexpected `story_id` attribute to Comment
Commenting on stories doesn't work with WSME >= 0.8 since that version of
WSME introduced a check on unexpected attributes in complex types. We
currently send an unexpected `story_id` attribute due to the way we pass
the story ID to the Angular $resource to decide the correct URL.

This commit changes the approach so that the story ID is passed only to
the Comment.$create() call rather than as an attribute of the object to
be created.

Change-Id: Ic6702d731a94f2468a3e66748538485a7546dcda