Link to the relevant story in comment notifications
Currently, the notification just says "$person left a comment $time"
which is entirely unhelpful and doesn't provide any kind of hint of
where the commment is. The intended behaviour is for a link to the
story to be displayed too, but that doesn't work as expected. This
patch fixes up the user_comment template to correctly link to the
story containing the comment.

Change-Id: I5f60a499c8460069fe87f04a5d47987d4e946c0b