Don't save default preference values
If a preference is being set to a default value due to it not being
set in the db, we shouldn't try to save the default value. This was
broken because the assumption in the comment was no longer true due
to Preference.refresh() being called on page load.

This change passes a flag when setting default preference values,
instead of trying to magically decide if we're setting defaults or
not.

Change-Id: I6c5f4ae8092fdcb486465f167510f641b4253df7