Make it clear that all fields are required
This is a quick and ugly fix. It literally adds '(required)' in the
placeholder text, wherever the field is required. In practice, that's
all fields. I've removed the red warnings that do this, because they
weren't consistent and I didn't want the user to be informed in a
different way depending on what field they were filling in.

I think this looks bad. I'd like a better solution, though I think it's
less confusing than the current state of affairs. I'm sending this for
review now so it's easier to see what the problem is!

Story: https://storyboard.openstack.org/#!/story/2000417

Change-Id: I097b3ed511ab5d3d4e5909cf22a3ee3dc7932101