Make it clear that all fields are required
This commit makes it obvious to a user that all relevant fields must be
filled in, and also makes story, project  and project group modals
consistent.

WIP because this doesn't yet work for multiple tasks/projects in
stories, or projects in project groups. There's enough here to
see if we want to use this approach in those fields, which are getting
a bit crowded.

Change-Id: Idf7f5e393d635686ed4dd631742adea277ef4493