Always go to Story Detail Page after adding Story
Before, creating a story from the project story list would just
refresh the list, rather than navigate to the new story. However,
storyboard would navigate to the new story when using the button
in the header, or the 'create story' button on the stories list page.
This was inconsistent and confusing, so this makes the behaviour the
same for all 'create story' buttons.

Story: https://storyboard.openstack.org/#!/story/2000075

Change-Id: I27391230717c82fb30237cf2714702de16da47a4