Logging in no longer causes page refresh.
This patch updates the LastLocation service to store data from the
UI-Router state, rather than the url, which allows us to navigate
to a state directly using UI-Router.

It also updates the SearchParamProvider to grab its data in the config
phase, before the UI-Router is initialized. This permits us to
forcefully remove any query parameters passed to our application in
the URL - as happens during the OAuth exchange - before the router
is aware of them, and thus circumvents an infinite digest loop.

The end result is that StoryBoard no longer has to reload itself
after a user is successfully authorized.

Note: The URL class used is not supported by PhantomJS. As such,
PhantomJS has been removed from our karma.conf, and Chrome has
been added.

Change-Id: Iae113544ad35708f63c5636189f0c88990eab203