Use improved worklists/boards API
This patch reduces the ridiculous number of GET requests done by
the worklists and boards pages to just a single GET which returns
a board/worklist and its contents. This makes loading of boards
appear much smoother and removes the "cards pop into existence
one by one" problem.

This patch doesn't change how permissions are handled, so they
still requre additional GETs. This should be improved in a later
patch if necessary.

This requires https://review.openstack.org/271408 in order for the
result of `GET /v1/boards/:id` and `GET /v1/worklists/:id` to be of
the newly expected form.

Change-Id: I357a311073bb94ab53cac8668399cca55c109ca5