Make header 'quick navigate' and 'search' clearer
There's a thing in the header that looks like a search bar. But its
*main* purpose is to let a user quickly browse to a resource by typing
in its name. It does also offer a search, but that search is very basic
(it can't be filtered). This patch renames the bar from 'Search' to
'Jump to...', and removes the search icon from the bar, so that
users aren't misled into thinking it offers a sophisticated search.
There is still the search option in the dropdown for users who want the
basic search.

Change-Id: I18909ae09834bae4af2a60363581e7beb0b02b7e