WIP for boards page
This is a page to view all boards/worklists. WIP because I'm not sure
it reacts correctly board/worklist permissions, and the code is a bit
inconsistent.

This depends on search icon patches to not make the sidebar get too
full of icons.

Depends on: https://review.openstack.org/#/c/278987/

Depends on: https://review.openstack.org/#/c/279204/

Change-Id: I12def9cb762e8270ce918ee1da961347836b3576