WIP 'All Boards and Worklists' Page
This is a page to view all boards/worklists. It doesn't yet filter
results based on criteria, though the skeleton of the html shows
how this will look.

Currently, a user can only find a board/worklist for which they are
explicitly listed as an owner or user, or for which they know the url.
This makes it hard for users to know that the feature exists, or to
see or demo examples.

Hopefully, the 'move search icon' patch can be merged soon after this,
to stop the sidebar from getting too full of icons.

The search does not work yet, hence wip.

Change-Id: I12def9cb762e8270ce918ee1da961347836b3576