'All Boards and Worklists' Page
This is a page to view all boards/worklists. It doesn't yet filter
results based on criteria; I think this should be an improvement in
a later patch.

Currently, a user can only find a board/worklist for which they are
explicitly listed as an owner or user, or for which they know the url.
This makes it hard for users to know that the feature exists, or to
see or demo examples.

This depends on the 'move search icon' patch to not make the sidebar
get too full of icons.

Depends on: https://review.openstack.org/#/c/279204/

Change-Id: I12def9cb762e8270ce918ee1da961347836b3576