WIP 'All Boards and Worklists' Page
This is a page to view all boards/worklists. It doesn't yet filter
results based on criteria; I think this should be an improvement in
a later patch.

Currently, a user can only find a board/worklist for which they are
explicitly listed as an owner or user, or for which they know the url.
This makes it hard for users to know that the feature exists, or to
see or demo examples.

Hopefully, the 'move search icon' patch can be merged soon after this,
to stop the sidebar from getting too full of icons.

The latest change adds pagination so that the page doesn't load too
slowly. Since I don't yet know how to do pagination, this is WIP;
it's mainly code tweaked from the project-story-list pagination. Thus
the page number doesn't yet display properly, the number of results
per page is hardcoded as '10', and there are pointless variables
hanging around.

Change-Id: I12def9cb762e8270ce918ee1da961347836b3576