Add rudimentary comment "permalink" functionality
This commit adds a button to comments which directs you to a link that
will automatically scroll the page to the comment.

The button won't work for comments past the first page, since it relies
on an id attribute in the comment's div element (which won't exist for
comment pages past the first on page load).

We could solve that by either:

a) Make the permalink point to a special "comments" view which displays
   all the comments on a story.
b) Split the Timeline Events out from comments, and load all comments on
   a story by default.

Change-Id: Iacf43c81457b018a30a2f1a410cfa50bea972fef