WIP Improve 'Tasks Assigned to Me'
Currently, we have 'stories assigned to me' on the dashboard, which
lists any active story containing a task assigned to the user. This
means there can be noise; a story may be active but the user's *tasks*
might all be merged or invalid (and some other user has an active task
in the story).

This patch changes the dashboard to display 'tasks assigned to me',
listing only a user's tasks that are todo, in progress, or in review.
It is wip because: 1) It does one GET per task-status, whereas it
should GET all a user's tasks and then filter. 2) The html is too
repetitive because of that. 3) the html is a bit inconsistent, and the
different task status labels are all the same colour.

Change-Id: I81a6504fd26e56a8dff0559816068d711e3c2c4b