Remove 'Subscriptions' button from header dropdown
There's a button to access all subscriptions in a dropdown menu
from the header. There's already a button to access all subscriptions
in a foldout menu from the sidebar. So, both these buttons require the
same number of clicks to access, and I feel the location of this one
is slightly less obvious than the other (the other one is grouped with
all 'this is relevant to me' things), and it's not clear why
subscriptions are on this menu when other resources are not. So I'd
rather remove it.

Change-Id: I81289f85583a7dcd2f79429b99833a60994b8b59