Add UI for making stories private to a list of people
This commit allows stories to be made private using the web UI. It
also allows you to define and modify a list of people that the story
is private to (i.e. they and only they can see it). Stories can also
be made public in the same way.

Private stories have a red warning box at the top to help indicate
the privacy quickly.

Change-Id: I64177a0708e41b39cbed8ac503630d613b374800
Depends-on: Ibd99032611ba1fd82de706e4d25acb1e2b98808c