WIP: Complex priorities UI in stories
This WIP adds a list of worklists which contain the story to the
story detail view. With https://review.openstack.org/#/c/310287/
it will show a list of worklists which contain the story or any
task within the story.

It also shows "100/120", which needs to be replaced by the position
of the relevant item in the worklist.

Change-Id: I399ac5017db9644ceddcaaba48775e6e2d9d0158