Complex priorities UI in stories
This commit adds a list of worklists that the current user is
subscribed to which contain the story or any task within the story
to the story view. This is a basic way of displaying complex
priorities for the story.

This commit also removes the priority dropdown, since we don't want
to support simple/global priority in StoryBoard really.

Story: 329
Task: 2987
Change-Id: I399ac5017db9644ceddcaaba48775e6e2d9d0158