Complex priorities UI in stories
This commit adds a list of worklists which contain the story to the
story detail view. With https://review.openstack.org/#/c/310287/
it will show a list of worklists which contain the story or any
task within the story. This is a basic way of displaying complex
priorities for the story.

This commit also removes the priority dropdown, since we don't want
to support simple/global priority in StoryBoard really.

Change-Id: I399ac5017db9644ceddcaaba48775e6e2d9d0158