Make task notes button clearer
It's hard to see that the task notes icon is a sticky note, especially
when it is black on a dark background. I constantly see people
directing others to it as "the dark red button", and think making it
easier to tell what the button is should alleviate this issue somewhat.

Change-Id: I4f46bc498df99905d7047c267a8cc8a99377325f