An alternative approach for making required fields clear
This patch denotes required fields with an asterisk,
which is a widely used shorthand. At the moment, storyboard
uses help text to denote required fields, along with crosses
and ticks (although it doesn't always use them). That can
get crowded when there are many required fields. We may want
to combine elements from both approaches. Either way,
storyboard should be consistent. This patch is an example
of a less-cluttered, but less informative approach, so we
can pick what to prioritize.

Change-Id: I87c3201bf7738179413b10be6de3db116bc79041