Always fall back to default config values
Currently, default values are set when `GET /config.json` fails.
This means that when there is a config file with some missing
(and required) keys, the constants resulting from those missing
keys are never set.

This commit makes the default config be added to the contents of
config.json, to set default values for any missing keys.

Change-Id: I3cfd33645d5b7f1a16851278e90b0b5c22d91cda